linkedin-prospector/requirements.txt
requests>=2.28.0
anthropic>=0.39.0
python-dotenv>=1.0.0
